@@ -30,16 +30,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^17.0.6",
-    "@angular/animations": "^17.0.6",
-    "@angular/common": "^17.0.6",
-    "@angular/compiler": "^17.0.6",
-    "@angular/core": "^17.0.6",
-    "@angular/forms": "^17.0.6",
-    "@angular/platform-browser": "^17.0.6",
-    "@angular/platform-browser-dynamic": "^17.0.6",
-    "@angular/platform-server": "^17.0.6",
-    "@angular/router": "^17.0.6",
+    "@angular-devkit/build-angular": "^18.2.5",
+    "@angular/animations": "^18.2.5",
+    "@angular/common": "^18.2.5",
+    "@angular/compiler": "^18.2.5",
+    "@angular/core": "^18.2.5",
+    "@angular/forms": "^18.2.5",
+    "@angular/platform-browser": "^18.2.5",
+    "@angular/platform-browser-dynamic": "^18.2.5",
+    "@angular/platform-server": "^18.2.5",
+    "@angular/router": "^18.2.5",
     "@angular/youtube-player": "^17.0.3",
     "@capacitor/android": "^6.0.0",
     "@capacitor/angular": "^2.0.3",
@@ -70,8 +70,8 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular/cli": "~17.0.6",
-    "@angular/compiler-cli": "^17.0.6",
+    "@angular/cli": "~18.2.5",
+    "@angular/compiler-cli": "^18.2.5",
     "@capacitor/assets": "^3.0.4",
     "@capacitor/cli": "^6.0.0",
     "@ionic/angular-toolkit": "^10.0.0",
@@ -86,6 +86,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.5.4"
   }
 }
